Add Netlify API functionality with TypeScript functions and homepage
- Add TypeScript Netlify Functions for proxy and countries APIs
- Create responsive homepage with API documentation and testing
- Update netlify.toml configuration for proper deployment
- Replace JavaScript functions with TypeScript for better Netlify support
