+++ functions/countries.ts
@@ -1,7 +1,20 @@
-const fs = require('fs');
-const path = require('path');
+// 直接内嵌代理数据，避免文件读取问题
+const proxyData = {
+  "proxy_list": [
+    {"host": "115.114.77.133:9090", "country": "IN", "quality": 4},
+    {"host": "41.223.119.156:3128", "country": "ZM", "quality": 4},
+    {"host": "52.188.28.218:3128", "country": "US", "quality": 4},
+    {"host": "128.199.120.45:9090", "country": "SG", "quality": 4},
+    {"host": "120.92.212.16:7890", "country": "CN", "quality": 4},
+    {"host": "211.230.49.122:3128", "country": "KR", "quality": 4},
+    {"host": "200.174.198.32:8888", "country": "BR", "quality": 4},
+    {"host": "39.185.41.193:5911", "country": "CN", "quality": 4},
+    {"host": "150.107.140.238:3128", "country": "ID", "quality": 4},
+    {"host": "38.183.183.114:999", "country": "PE", "quality": 4}
+  ]
+};
 
-exports.handler = async (event, context) => {
+export const handler = async (event: any, context: any) => {
   try {
     // 设置CORS头
     const headers = {
@@ -29,43 +42,16 @@ exports.handler = async (event, context) => {
       };
     }
 
-    // 读取代理列表
-    const proxyListPath = path.join(__dirname, '..', 'pxy.json');
-    const proxyData = fs.readFileSync(proxyListPath, 'utf8');
-    const proxies = JSON.parse(proxyData);
-
-    // 获取查询参数
-    const { country, limit, random } = event.queryStringParameters || {};
-
-    let filteredProxies = proxies.proxy_list;
-
-    // 按国家过滤
-    if (country) {
-      filteredProxies = filteredProxies.filter(proxy => 
-        proxy.country.toLowerCase() === country.toLowerCase()
-      );
-    }
-
-    // 随机排序
-    if (random === 'true') {
-      filteredProxies = filteredProxies.sort(() => Math.random() - 0.5);
-    }
-
-    // 限制数量
-    if (limit) {
-      const limitNum = parseInt(limit);
-      if (!isNaN(limitNum) && limitNum > 0) {
-        filteredProxies = filteredProxies.slice(0, limitNum);
-      }
-    }
+    // 提取所有不重复的国家代码
+    const countries = [...new Set(proxyData.proxy_list.map((proxy: any) => proxy.country))].sort();
 
     return {
       statusCode: 200,
       headers,
       body: JSON.stringify({
         success: true,
-        count: filteredProxies.length,
-        proxies: filteredProxies
+        count: countries.length,
+        countries: countries
       })
     };
